requirements.txt
streamlit==1.28.0
pandas==2.0.3
biopython==1.81
requests==2.31.0